chore: convert app installation to typescript
